backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sock==0.7.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
